refactor: move few components from ui to components directory
