+++ src/components/currencyList/CurrencyList.module.scss
