+++ src/components/currencyList/CurrencyList.tsx
