+++ src/components/currencySearch/CurrencySearch.tsx
