+++ src/pages/home/HomePage.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { useDispatch } from 'react-redux';
 
-import { CurrencyList } from '@/components/ui/currencyList/CurrencyList';
+import { CurrencyList } from '@/components/currencyList/CurrencyList';
 import { stocks } from '@/constants/currencies/stocks';
 import { getCurrencies } from '@/services/currency/getCurrencies';
 import { updateCurrencies } from '@/store/slices/currencySlice';
